refactor(provider): centralize state reconstruction
